backend-service/requirements.txt
Django>=3.0,<4.0
gunicorn==20.1.0
django-cors-headers 
djangorestframework 
pillow 
braintree 